requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
cohere
pypdf
faiss-cpu
sentence-transformers
langchain
langchain-community
langchain-huggingface
